+++ class_to_string/lib/src/utils.dart
@@ -1,0 +1,16 @@
+extension WriteValueStringBuggerExtension on StringBuffer {
+  void writeValue(Object? value) {
+    switch (value) {
+      case String():
+        write("'");
+        write(value);
+        write("'");
+
+      case null:
+      case num():
+      case bool():
+      default:
+        write(value);
+    }
+  }
+}