+++ mek_data_class_generator/lib/builder.dart
@@ -1,16 +1,15 @@
 import 'package:build/build.dart';
 import 'package:mek_data_class_generator/mek_data_class_generator.dart';
-import 'package:mek_data_class_generator/src/configs.dart';
+import 'package:mek_data_class_generator/src/configs/options.dart';
 import 'package:source_gen/source_gen.dart';
 
 /// Dart entry point for building a DataClass generated files
-Builder dataClass(BuilderOptions options) {
-  final config = Config.fromJson(options.config);
-  final pageWidth = config.pageWidth;
-  if (pageWidth != null) {
-    log.warning('mek_data_class_generator: The "page_width: $pageWidth" option has been replaced '
-        'by the "formatter: { page_width: 80 }" option in the "analysis_options.yaml" file');
-  }
+Builder dataClass(BuilderOptions buildOptions) {
+  final options = Options.fromJson(buildOptions.config);
 
-  return SharedPartBuilder([DataClassGenerator(config: config)], 'data_class');
+  return SharedPartBuilder(
+    [DataClassGenerator(options: options)],
+    'data_class',
+    allowSyntaxErrors: true,
+  );
 }