+++ class_to_string/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.1.0
+- feat: The output is formatted the same as some dart code
+
 ## 1.0.0
 - feat: `ClassToString.add` method now stringify a `null` values [#17](https://github.com/BreX900/data_class/issues/17)